+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **[BREAKING]** Do not replace initial values when form was successfully submitted
 - Added `onLoad` option in `useForm()`
 - Added attributes `loadError`, `loaded` and `loading` returned by `useForm()`
+- Fixed `form.errors` returning errors with `undefined`
 
 ## v3.0.9